requirements.txt

    # Django Framework and Extensions
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-templated-mail==1.1.1
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
django-apscheduler==0.6.2
django-celery-beat==2.5.0
django-celery-results==2.5.1


    # Database Clients
mysqlclient==2.2.0
psycopg==3.1.12
psycopg2==2.9.9
redis==5.0.1
flower==2.0.1



    # Authentication and Social Login
oauthlib==3.2.2
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.4.2


    # API Documentation
drf-yasg==1.21.7


    # Security and Data Protection
cryptography==41.0.4
defusedxml==0.8.0rc2


    # Utilities and Helpers
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
idna==3.4
pycparser==2.21
python-decouple==3.8
requests==2.31.0
urllib3==2.0.7
psycopg==3.1.12
psycopg2==2.9.9
Pillow==10.1.0
celery==5.3.6


    # Timezone Data
pytz==2023.3.post1
tzdata==2023.3


    # ASGI Handling
asgiref==3.7.2


    # Image Processing
Pillow==10.1.0


    # SQL Parsing
sqlparse==0.4.4

